Connect supabase with our app and make a first fetch data test
From the Supabase client, we can now create queries with the from method. And so then we specify the name of the table and then the fields that we want. Here we want basically all of them. And so this returns a promise which we then await and the result of that gives us the data and a possible error. If there is no error, then we simply return that data.
And so now we are ready to use this service anywhere in our application.
We activated row level security, that's why it's not problem that we're exposing our supabase key on the client. If any malicious user tries to hack our database - he won't be able to do it, exactly because we activated the row level security. Now anyone who has this key can only do whatever we allowed in the row level security policies. And so right now the only thing that we allow is to read data from our four (for now) tables. And that's why it is completely safe to expose this key here to the client.