Set up two storage buckets in the database where we can upload large files to.
We went to 'storage' and created two buckets - 'avatars' and 'cabin-images' and uploaded the cabin images in 'cabin-images', took/copied the URL, placed it here, as img src, and it is already rendered in the page.
We put the image against img (which was set to be text) in the database and now it URL is shown against the image key in the browser when we log it.