Build a simple modal window component using React's portal feature so that we can then convert this modal to a compound component later
In our cabins page we actually want the form to add a new cabin to appear in a modal window, so not just rendered on the page as it was until now, but we want basically a new window to open on top of this and then display our form in there.
For that we are now building that reusable modal component.