Converting the Modal Component to a Compound Component
The reason why that is necessary is that this modal that we have built is really not ideal when it comes to the state management and to the way in which we actually render this modal.
It shouldn't be a task of the AddCabin component here to track whether right now the modal should be displayed or not. So instead, the modal component itself should actually know whether it is currently open or not, and so it should keep this state internally.
Also fix the styling of the modal window by removing the redundant padding.