Building a reusable context menu
Give each of the cabin rows a button that we can click. And when we click that button, a small context menu appears, which ultimately contain the three actions - for duplicating, editing, and deleting.

We actually wrap everything in CabinRow.jsx, so all the rows inside those menus so that those menus can then keep track which of the menus are actually opened at the time.

Do also so when we click outside the context menu, the menu to close, using the useOutsideClick hook.