Implement the bookings filtering
This time we did it in a fundamentally way than what we did with the cabins, where the operations happened already on the client side.
Now with the bookings we actually want to do it on the server side, on the API side of the data. Or in other words, if i want to filter for example, for "checked out", then i want the API to really only send me all the bookings that have the "checked out" status. So i don't want to receive all of them, and then just filter them here on the client side. But instead, really only that filtered data should get downloaded from supabase.