Implement checking in in a booking.
Bookings may not have been paid yet on guest arrival. Therefore, on check in, users need to accept payment (outside the app), and then confirm that payment has been received (inside the app). So we're doing this, creating a new React Query custom hook.
Basically checking in means to change the status from unconfirmed to checked in. Being checked in means that the booking status is indeed set to checked-in and also is paid will need to be set to true (making sure that the employee has actually received the payment). For that reason we are now using a mutation where we update a booking by setting these two fields -> the status to checkin and the isPaid to true.