Implement the option for employees to check out a guest once they leave the hotel
Doing this by using React Query custom hook and fix small bug.

Information about the bug - the problem were that when we are on page three in the bookings, for example, and we selected "unconfirmed", we were still trying to request data for page number three even though with the status of unconfirmed, there are no three pages of data.
This problem is now fixed.