Fix a small bug
If we are not logged in and try to visit the root URL (localhost:5173), it redirects us to the login and we also then get this ["user"] in our cache, which is just null right now. The reason for this is if we try to access the root URL, basically what we're accessing is this dashboard here (look in App.jsx). But this dasboard is of course inside the protected route. And if we are not authenticated, we're redirected to the login.
In fact, the bug is in this line of code in useLogin.js:       queryClient.setQueriesData(["user"], user);
Here it should not be setQueriesData, but setQueryData. And also the user that we have here is wrong.
The problem is that in that way if we log in successfully, this isAuthenticated state in useUser will still not change to true, because we have a wrong function in useLogin.js (setQueriesData). And so therefore then we are not redirected to the content itself, so basically to the dashboard.
But in that way the user that we are adding to React chache is actually wrong. Because we have an object which contains both the session and the user (the instructor logs it in the video), and this is not what we want.
We fixed this by putting setQueryData instead of setQueriesData and user.user instead of only user.