Setup react error boundary to show user friendly message in case of errors while rendering
Install it with npm i react-error-boundary
And so this package now all it does is to give us an error boundary component where we can pass in a fallback and also a function to reset the application, so whenever an error has occurred.
To use this, we basically wrap our entire application into that error boundary.