Fix two last small bugs
In the Cabins page if we clicked on the three vertical points and open the additional menu - the menu didn't close when we click on the same three vertical points. Now this is fixed.

Also, if we tried to open a booking that doesn't exist, it was rendering page from the the error boundary component which we built in the previous lecture. But this wasn't so good, because what should happening is not an error, but just the information that this booking couldn't be found. We fixed this. Now in cases like this the Empty component is shown.